% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aiptw.R
\name{aiptw}
\alias{aiptw}
\title{Augmented Inverse Probability of Treatment Weighted Estimation}
\usage{
aiptw(data, cum.g, Ynodes, Anodes, Cnodes = NULL, abar, Qform,
  SL.library = NULL, family = "quasibinomial", stratify = TRUE)
}
\arguments{
\item{data}{data frame following the time-ordering of the nodes.}

\item{cum.g}{a matrix of the cumulative probabilities of treatment (and being uncensored) given the parents.}

\item{Ynodes}{column names or indicies in \code{data} of outcome nodes.}

\item{Anodes}{column names or indicies in \code{data} of treatment nodes.}

\item{Cnodes}{column names or indicies in \code{data} of censoring nodes.}

\item{abar}{binary vector (numAnodes x 1) of counterfactual treatment}

\item{Qform}{character vector of regression formulas for Q.}

\item{SL.library}{optional character vector of libraries to pass to SuperLearner. NULL indicates glm should be called instead of SuperLearner.}

\item{stratify}{if \code{TRUE} condition on following \code{abar} when estimating Q and g. If \code{FALSE}, pool over all subjects.}
}
\value{
\code{BangRobinsDR} returns a list of items as an object of class \code{aiptw}, which include
\itemize{
	\item {The estimate of the parameter value under the intervention \code{abar}}
	\item {The empirical influence function for the point estimate}
	\item {The time-specific empirical influence functions}
	\item {The conditional expectation fits for \code{Qbar}}
	\item {The call to the function}
}
}
\description{
\code{aiptw} Estimates the parameter of interest (E[Y_d]) by directly solving the efficient influence function as an estimating equation.
}
\details{
Please refer to van der Laan and Gruben (2007) for details regarding derivation of the efficient influence function, which this estimator solves.
}

