% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/icedr.R
\name{icedr}
\alias{icedr}
\title{Double Robust Sequential Regression Estimation}
\usage{
icedr(data, Ynodes, Anodes, Cnodes = NULL, abar, cum.g, Qform,
  SL.library = NULL, family = "quasibinomial", g.weight = FALSE,
  stratify = TRUE)
}
\arguments{
\item{data}{data frame following the time-ordering of the nodes.}

\item{Ynodes}{column names or indicies in \code{data} of outcome nodes.}

\item{Anodes}{column names or indicies in \code{data} of treatment nodes.}

\item{Cnodes}{column names or indicies in \code{data} of censoring nodes.}

\item{abar}{binary vector (numAnodes x 1) of counterfactual treatment}

\item{cum.g}{a matrix of the cumulative probabilities of treatment (and being uncensored) given the parents.}

\item{Qform}{character vector of regression formulas for \code{Qbar}.}

\item{SL.library}{optional character vector of libraries to pass to SuperLearner. NULL indicates glm should be called instead of SuperLearner.}

\item{family}{Currently allows \code{gaussian} or \code{binomial} to describe the error distribution. Link function information will be ignored.}

\item{g.weight}{if \code{TRUE}, then use the g fits as weights in the \code{Qbar} fit, rather than as a covariate.}

\item{stratify}{if \code{TRUE} condition on following \code{abar} when estimating \code{Qbar}. If \code{FALSE}, pool over all subjects.}
}
\value{
\code{icedr} returns a list of items as an object of class \code{icedr}, which include
\itemize{
	\item {The estimate of the parameter value under the intervention \code{abar}}
	\item {A matrix of the condition expectation fit \code{Qbar} under \code{abar} for each time point.}
	\item {The conditional expectation fits for \code{Qbar}}
	\item {The call to the function}
}
}
\description{
\code{icedr} Estimates the parameter of interest (E[Y_d]) by using the sequential regression approach as presented by Bang and Robins (2005). This function only works in the setting where the outcome is binary or the outcome is continuous with a censoring intervention
}

